Importando versión upstream de iceweasel 22

+++ layout/tools/reftest/remotereftest.py
@@ -1,45 +1,12 @@
-# ***** BEGIN LICENSE BLOCK *****
-# Version: MPL 1.1/GPL 2.0/LGPL 2.1
-#
-# The contents of this file are subject to the Mozilla Public License Version
-# 1.1 (the "License"); you may not use this file except in compliance with
-# the License. You may obtain a copy of the License at
-# http://www.mozilla.org/MPL/
-#
-# Software distributed under the License is distributed on an "AS IS" basis,
-# WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
-# for the specific language governing rights and limitations under the
-# License.
-#
-# The Original Code is mozilla.org code.
-#
-# The Initial Developer of the Original Code is Joel Maher.
-#
-# Portions created by the Initial Developer are Copyright (C) 2010
-# the Initial Developer. All Rights Reserved.
-#
-# Contributor(s):
-#  Joel Maher <joel.maher@gmail.com> (Original Developer)
-#  Clint Talbert <cmtalbert@gmail.com>
-#
-# Alternatively, the contents of this file may be used under the terms of
-# either the GNU General Public License Version 2 or later (the "GPL"), or
-# the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
-# in which case the provisions of the GPL or the LGPL are applicable instead
-# of those above. If you wish to allow use of your version of this file only
-# under the terms of either the GPL or the LGPL, and not to allow others to
-# use your version of this file under the terms of the MPL, indicate your
-# decision by deleting the provisions above and replace them with the notice
-# and other provisions required by the GPL or the LGPL. If you do not delete
-# the provisions above, a recipient may use your version of this file under
-# the terms of any one of the MPL, the GPL or the LGPL.
-#
-# ***** END LICENSE BLOCK *****
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import sys
 import os
 import time
 import tempfile
+import traceback
 
 # We need to know our current directory so that we can serve our test files from it.
 SCRIPT_DIRECTORY = os.path.abspath(os.path.realpath(os.path.dirname(sys.argv[0])))
@@ -48,7 +15,7 @@ from runreftest import RefTest
 from runreftest import ReftestOptions
 from automation import Automation
 import devicemanager, devicemanagerADB, devicemanagerSUT
-from remoteautomation import RemoteAutomation
+from remoteautomation import RemoteAutomation, fennecLogcatFilters
 
 class RemoteOptions(ReftestOptions):
     def __init__(self, automation):
@@ -57,7 +24,6 @@ class RemoteOptions(ReftestOptions):
         defaults = {}
         defaults["logFile"] = "reftest.log"
         # app, xrePath and utilityPath variables are set in main function
-        defaults["remoteTestRoot"] = None
         defaults["app"] = ""
         defaults["xrePath"] = ""
         defaults["utilityPath"] = ""
@@ -85,7 +51,7 @@ class RemoteOptions(ReftestOptions):
         self.add_option("--remote-webserver", action="store",
                     type = "string", dest = "remoteWebServer",
                     help = "IP Address of the webserver hosting the reftest content")
-        defaults["remoteWebServer"] = automation.getLanIp() 
+        defaults["remoteWebServer"] = automation.getLanIp()
 
         self.add_option("--http-port", action = "store",
                     type = "string", dest = "httpPort",
@@ -111,10 +77,19 @@ class RemoteOptions(ReftestOptions):
                     help = "name of the pidfile to generate")
         defaults["pidFile"] = ""
 
+        self.add_option("--bootstrap", action="store_true", dest = "bootstrap",
+                    help = "test with a bootstrap addon required for native Fennec")
+        defaults["bootstrap"] = False
+
         self.add_option("--dm_trans", action="store",
                     type = "string", dest = "dm_trans",
                     help = "the transport to use to communicate with device: [adb|sut]; default=sut")
         defaults["dm_trans"] = "sut"
+
+        self.add_option("--remoteTestRoot", action = "store",
+                    type = "string", dest = "remoteTestRoot",
+                    help = "remote directory to use as test root (eg. /mnt/sdcard/tests or /data/local/tests)")
+        defaults["remoteTestRoot"] = None
 
         defaults["localLogName"] = None
 
@@ -122,7 +97,8 @@ class RemoteOptions(ReftestOptions):
 
     def verifyRemoteOptions(self, options):
         # Ensure our defaults are set properly for everything we can infer
-        options.remoteTestRoot = self._automation._devicemanager.getDeviceRoot() + '/reftest'
+        if not options.remoteTestRoot:
+            options.remoteTestRoot = self._automation._devicemanager.getDeviceRoot() + '/reftest'
         options.remoteProfile = options.remoteTestRoot + "/profile"
 
         # Verify that our remotewebserver is set properly
@@ -144,23 +120,23 @@ class RemoteOptions(ReftestOptions):
             # Neither remoteAppPath nor app are set -- error
             print "ERROR: You must specify either appPath or app"
             return None
-        
+
         if (options.xrePath == None):
             print "ERROR: You must specify the path to the controller xre directory"
             return None
         else:
             # Ensure xrepath is a full path
             options.xrePath = os.path.abspath(options.xrePath)
-        
+
         # Default to <deviceroot>/reftest/reftest.log
         if (options.remoteLogFile == None):
             options.remoteLogFile = 'reftest.log'
 
         options.localLogName = options.remoteLogFile
         options.remoteLogFile = options.remoteTestRoot + '/' + options.remoteLogFile
-         
+
         # Ensure that the options.logfile (which the base class uses) is set to
-        # the remote setting when running remote. Also, if the user set the 
+        # the remote setting when running remote. Also, if the user set the
         # log file name there, use that instead of reusing the remotelogfile as above.
         if (options.logFile):
             # If the user specified a local logfile name use that
@@ -198,7 +174,7 @@ class ReftestServer:
 
     def start(self):
         "Run the Refest server, returning the process ID of the server."
-          
+
         env = self._automation.environment(xrePath = self._xrePath)
         env["XPCOM_DEBUG_BREAK"] = "warn"
         if self._automation.IS_WIN32:
@@ -207,17 +183,25 @@ class ReftestServer:
         args = ["-g", self._xrePath,
                 "-v", "170",
                 "-f", os.path.join(self.scriptDir, "reftest/components/httpd.js"),
-                "-e", "const _PROFILE_PATH = '%(profile)s';const _SERVER_PORT = '%(port)s'; const _SERVER_ADDR ='%(server)s';" % 
+                "-e", "const _PROFILE_PATH = '%(profile)s';const _SERVER_PORT = '%(port)s'; const _SERVER_ADDR ='%(server)s';" %
                        {"profile" : self._profileDir.replace('\\', '\\\\'), "port" : self.httpPort, "server" : self.webServer },
                 "-f", os.path.join(self.scriptDir, "server.js")]
 
         xpcshell = os.path.join(self._utilityPath,
                                 "xpcshell" + self._automation.BIN_SUFFIX)
+
+        if not os.access(xpcshell, os.F_OK):
+            raise Exception('xpcshell not found at %s' % xpcshell)
+        if self._automation.elf_arm(xpcshell):
+            raise Exception('xpcshell at %s is an ARM binary; please use '
+                            'the --utility-path argument to specify the path '
+                            'to a desktop version.' % xpcshell)
+
         self._process = self._automation.Process([xpcshell] + args, env = env)
         pid = self._process.pid
         if pid < 0:
             print "Error starting server."
-            sys.exit(2)
+            return 2
         self._automation.log.info("INFO | remotereftests.py | Server pid: %d", pid)
 
         if (self.pidFile != ""):
@@ -238,19 +222,20 @@ class ReftestServer:
         else:
             print "Timed out while waiting for server startup."
             self.stop()
-            sys.exit(1)
+            return 1
 
     def stop(self):
-        try:
-            c = urllib2.urlopen(self.shutdownURL)
-            c.read()
-            c.close()
+        if hasattr(self, '_process'):
+            try:
+                c = urllib2.urlopen(self.shutdownURL)
+                c.read()
+                c.close()
 
-            rtncode = self._process.poll()
-            if (rtncode == None):
-                self._process.terminate()
-        except:
-            self._process.kill()
+                rtncode = self._process.poll()
+                if (rtncode == None):
+                    self._process.terminate()
+            except:
+                self._process.kill()
 
 class RemoteReftest(RefTest):
     remoteApp = ''
@@ -290,85 +275,101 @@ class RemoteReftest(RefTest):
         localAutomation.UNIXISH = False
         hostos = sys.platform
         if (hostos == 'mac' or  hostos == 'darwin'):
-          localAutomation.IS_MAC = True
+            localAutomation.IS_MAC = True
         elif (hostos == 'linux' or hostos == 'linux2'):
-          localAutomation.IS_LINUX = True
-          localAutomation.UNIXISH = True
+            localAutomation.IS_LINUX = True
+            localAutomation.UNIXISH = True
         elif (hostos == 'win32' or hostos == 'win64'):
-          localAutomation.BIN_SUFFIX = ".exe"
-          localAutomation.IS_WIN32 = True
+            localAutomation.BIN_SUFFIX = ".exe"
+            localAutomation.IS_WIN32 = True
 
         paths = [options.xrePath, localAutomation.DIST_BIN, self.automation._product, os.path.join('..', self.automation._product)]
         options.xrePath = self.findPath(paths)
         if options.xrePath == None:
             print "ERROR: unable to find xulrunner path for %s, please specify with --xre-path" % (os.name)
-            sys.exit(1)
+            return 1
         paths.append("bin")
         paths.append(os.path.join("..", "bin"))
 
         xpcshell = "xpcshell"
         if (os.name == "nt"):
             xpcshell += ".exe"
-      
+
         if (options.utilityPath):
             paths.insert(0, options.utilityPath)
         options.utilityPath = self.findPath(paths, xpcshell)
         if options.utilityPath == None:
             print "ERROR: unable to find utility path for %s, please specify with --utility-path" % (os.name)
-            sys.exit(1)
+            return 1
 
         options.serverProfilePath = tempfile.mkdtemp()
         self.server = ReftestServer(localAutomation, options, self.scriptDir)
-        self.server.start()
+        retVal = self.server.start()
+        if retVal:
+            return retVal
+        retVal = self.server.ensureReady(self.SERVER_STARTUP_TIMEOUT)
+        if retVal:
+            return retVal
 
-        self.server.ensureReady(self.SERVER_STARTUP_TIMEOUT)
         options.xrePath = remoteXrePath
         options.utilityPath = remoteUtilityPath
-         
+        return 0
+
     def stopWebServer(self, options):
         self.server.stop()
 
-    def createReftestProfile(self, options, profileDir):
-        RefTest.createReftestProfile(self, options, profileDir, server=options.remoteWebServer)
+    def createReftestProfile(self, options, profileDir, reftestlist):
+        RefTest.createReftestProfile(self, options, profileDir, reftestlist, server=options.remoteWebServer)
+
+        # Turn off the locale picker screen
+        fhandle = open(os.path.join(profileDir, "user.js"), 'a')
+        fhandle.write("""
+user_pref("browser.firstrun.show.localepicker", false);
+user_pref("font.size.inflation.emPerLine", 0);
+user_pref("font.size.inflation.minTwips", 0);
+user_pref("reftest.remote", true);
+// Set a future policy version to avoid the telemetry prompt.
+user_pref("toolkit.telemetry.prompted", 999);
+user_pref("toolkit.telemetry.notifiedOptOut", 999);
+user_pref("reftest.uri", "%s");
+user_pref("datareporting.policy.dataSubmissionPolicyBypassAcceptance", true);
+""" % reftestlist)
 
         #workaround for jsreftests.
         if options.enablePrivilege:
-          fhandle = open(os.path.join(profileDir, "user.js"), 'a')
-          fhandle.write("""
-user_pref("capability.principal.codebase.p2.granted", "UniversalPreferencesWrite UniversalXPConnect UniversalBrowserWrite UniversalPreferencesRead UniversalBrowserRead");
+            fhandle.write("""
+user_pref("capability.principal.codebase.p2.granted", "UniversalXPConnect");
 user_pref("capability.principal.codebase.p2.id", "http://%s:%s");
 """ % (options.remoteWebServer, options.httpPort))
-          fhandle.close()
 
-        if (self._devicemanager.pushDir(profileDir, options.remoteProfile) == None):
-            raise devicemanager.FileError("Failed to copy profiledir to device")
+        # Close the file
+        fhandle.close()
+
+        try:
+            self._devicemanager.pushDir(profileDir, options.remoteProfile)
+        except devicemanager.DMError:
+            print "Automation Error: Failed to copy profiledir to device"
+            raise
 
     def copyExtraFilesToProfile(self, options, profileDir):
         RefTest.copyExtraFilesToProfile(self, options, profileDir)
-        if (self._devicemanager.pushDir(profileDir, options.remoteProfile) == None):
-            raise devicemanager.FileError("Failed to copy extra files to device") 
-
-    def registerExtension(self, browserEnv, options, profileDir, extraArgs = ['-silent'] ):
-        self.automation.log.info("REFTEST INFO | runreftest.py | Performing extension manager registration: start.\n")
-        # Because our startProcess code doesn't return until fennec starts we just give it
-        # a maxTime of 20 secs before timing it out and ensuring it is dead.
-        # Besides registering the extension, this works around fennec bug 570027
-        status = self.automation.runApp(None, browserEnv, options.app, profileDir,
-                                   extraArgs,
-                                   utilityPath = options.utilityPath,
-                                   xrePath=options.xrePath,
-                                   symbolsPath=options.symbolsPath,
-                                   maxTime = 20)
-        # We don't care to call |processLeakLog()| for this step.
-        self.automation.log.info("\nREFTEST INFO | runreftest.py | Performing extension manager registration: end.")
+        try:
+            self._devicemanager.pushDir(profileDir, options.remoteProfile)
+        except devicemanager.DMError:
+            print "Automation Error: Failed to copy extra files to device"
+            raise
 
     def getManifestPath(self, path):
         return path
 
     def cleanup(self, profileDir):
         # Pull results back from device
-        if (self.remoteLogFile):
+        if self.remoteLogFile and \
+                self._devicemanager.fileExists(self.remoteLogFile):
             self._devicemanager.getFile(self.remoteLogFile, self.localLogName)
+        else:
+            print "WARNING: Unable to retrieve log file (%s) from remote " \
+                "device" % self.remoteLogFile
         self._devicemanager.removeDir(self.remoteProfile)
         self._devicemanager.removeDir(self.remoteTestRoot)
         RefTest.cleanup(self, profileDir)
@@ -379,23 +380,27 @@ user_pref("capability.principal.codebase.p2.id", "http://%s:%s");
             except:
                 print "Warning: cleaning up pidfile '%s' was unsuccessful from the test harness" % self.pidFile
 
-def main():
-    dm_none = devicemanagerADB.DeviceManagerADB(None, None)
-    automation = RemoteAutomation(dm_none)
+def main(args):
+    automation = RemoteAutomation(None)
     parser = RemoteOptions(automation)
     options, args = parser.parse_args()
 
     if (options.deviceIP == None):
         print "Error: you must provide a device IP to connect to via the --device option"
-        sys.exit(1)
+        return 1
 
-    if (options.dm_trans == "adb"):
-        if (options.deviceIP):
-            dm = devicemanagerADB.DeviceManagerADB(options.deviceIP, options.devicePort)
+    try:
+        if (options.dm_trans == "adb"):
+            if (options.deviceIP):
+                dm = devicemanagerADB.DeviceManagerADB(options.deviceIP, options.devicePort, deviceRoot=options.remoteTestRoot)
+            else:
+                dm = devicemanagerADB.DeviceManagerADB(None, None, deviceRoot=options.remoteTestRoot)
         else:
-            dm = dm_none
-    else:
-         dm = devicemanagerSUT.DeviceManagerSUT(options.deviceIP, options.devicePort)
+            dm = devicemanagerSUT.DeviceManagerSUT(options.deviceIP, options.devicePort, deviceRoot=options.remoteTestRoot)
+    except devicemanager.DMError:
+        print "Error: exception while initializing devicemanager.  Most likely the device is not in a testable state."
+        return 1
+
     automation.setDeviceManager(dm)
 
     if (options.remoteProductName != None):
@@ -405,7 +410,7 @@ def main():
     options = parser.verifyRemoteOptions(options)
     if (options == None):
         print "ERROR: Invalid options specified, use --help for a list of valid options"
-        sys.exit(1)
+        return 1
 
     if not options.ignoreWindowSize:
         parts = dm.getInfo('screen')['screen'][0].split()
@@ -413,15 +418,13 @@ def main():
         height = int(parts[1].split(':')[1])
         if (width < 1050 or height < 1050):
             print "ERROR: Invalid screen resolution %sx%s, please adjust to 1366x1050 or higher" % (width, height)
-            sys.exit(1)
+            return 1
 
     automation.setAppName(options.app)
     automation.setRemoteProfile(options.remoteProfile)
     automation.setRemoteLog(options.remoteLogFile)
     reftest = RemoteReftest(automation, dm, options, SCRIPT_DIRECTORY)
-
-    # Start the webserver
-    reftest.startWebServer(options)
+    options = parser.verifyCommonOptions(options, reftest)
 
     # Hack in a symbolic link for jsreftest
     os.system("ln -s ../jsreftest " + str(os.path.join(SCRIPT_DIRECTORY, "jsreftest")))
@@ -433,15 +436,45 @@ def main():
     elif os.path.exists(args[0]):
         manifestPath = os.path.abspath(args[0]).split(SCRIPT_DIRECTORY)[1].strip('/')
         manifest = "http://" + str(options.remoteWebServer) + ":" + str(options.httpPort) + "/" + manifestPath
+    else:
+        print "ERROR: Could not find test manifest '%s'" % manifest
+        return 1
+
+    # Start the webserver
+    retVal = reftest.startWebServer(options)
+    if retVal:
+        return retVal
 
     procName = options.app.split('/')[-1]
     if (dm.processExist(procName)):
-      dm.killProcess(procName)
+        dm.killProcess(procName)
+
+    print dm.getInfo()
 
 #an example manifest name to use on the cli
 #    manifest = "http://" + options.remoteWebServer + "/reftests/layout/reftests/reftest-sanity/reftest.list"
-    reftest.runTests(manifest, options)
+    retVal = 0
+    try:
+        cmdlineArgs = ["-reftest", manifest]
+        if options.bootstrap:
+            cmdlineArgs = []
+        dm.recordLogcat()
+        retVal = reftest.runTests(manifest, options, cmdlineArgs)
+    except:
+        print "Automation Error: Exception caught while running tests"
+        traceback.print_exc()
+        retVal = 1
+
     reftest.stopWebServer(options)
+    try:
+        logcat = dm.getLogcat(filterOutRegexps=fennecLogcatFilters)
+        print ''.join(logcat)
+        print dm.getInfo()
+    except devicemanager.DMError:
+        print "WARNING: Error getting device information at end of test"
+
+    return retVal
 
 if __name__ == "__main__":
-    main()
+    sys.exit(main(sys.argv[1:]))
+
